# backend/swaggerDoc.yaml
openapi: 3.0.3
info:
  title: Disaster Management API
  version: 1.0.0
servers:
  - url: http://localhost:5555/api
    description: development server
  - url: https://cdrp-1.onrender.com/api
    description: production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Incident:
      type: object
      required: [type, title, description, dateOccurred, location, status]
      properties:
        _id: { type: string }
        type: { type: string, minLength: 4 }
        title: { type: string, minLength: 4 }
        description: { type: string, minLength: 4, maxLength: 300 }
        dateOccurred: { type: string, format: date-time }
        location:
          type: object
          properties:
            type: { type: string, enum: [Point] }
            coordinates: { type: array, items: { type: number } }
        status:
          {
            type: string,
            enum:
              [pending, validated, assigned, in progress, critical, resolved],
            default: pending,
          }
        reports: { type: array, items: { type: string } }
        priority: { type: string, enum: [high, medium, low], default: medium }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Assignment:
      type: object
      required: [incident, organization, assignedBy]
      properties:
        _id: { type: string }
        incident: { type: string }
        organization: { type: string }
        assignedBy: { type: string }
        notes: { type: string, maxLength: 500 }
        priorityLevel:
          { type: string, enum: [high, medium, low], default: medium }
        status:
          {
            type: string,
            enum: [active, completed, cancelled],
            default: active,
          }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /incidents:
    get:
      tags: [Incidents]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 5 }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          {
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/Incident" },
                      },
                  },
              },
          }
    delete:
      tags: [Incidents]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { content: { application/json: {} } }
  /incidents/{id}:
    get:
      tags: [Incidents]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          {
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Incident" } },
              },
          }
    patch:
      tags: [Incidents]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string }
                title: { type: string }
                description: { type: string }
                status: { type: string }
                location:
                  {
                    type: object,
                    properties:
                      {
                        type: { type: string },
                        coordinates: { type: array, items: { type: number } },
                      },
                  }
                dateOccurred: { type: string, format: date-time }
      responses:
        "200":
          {
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Incident" } },
              },
          }
    delete:
      tags: [Incidents]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { content: { application/json: {} } }
  /incidents/images/{id}:
    get:
      tags: [Incidents]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          {
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items:
                          {
                            type: object,
                            properties:
                              {
                                url: { type: string, format: uri },
                                cloudinaryId: { type: string },
                              },
                          },
                      },
                  },
              },
          }
  /incidents/{id}/assign:
    post:
      tags: [Incidents]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                incidentId: { type: string }
                organizationId: { type: string }
                notes: { type: string }
                priorityLevel: { type: string }
      responses:
        "201":
          {
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Assignment" } },
              },
          }
  /incidents/{id}/assignments:
    get:
      tags: [Incidents]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          {
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/Assignment" },
                      },
                  },
              },
          }
...
  /assignments/incident/{incidentId}:
    get:
      tags: [Assignments]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: incidentId
          required: true
          schema: { type: string }
      responses:
        "200":
          {
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/Assignment" },
                      },
                  },
              },
          }
...